requirements.txt
flet~=0.28.3
openpyxl~=3.1.5
pandas~=2.3.0
pillow~=11.3.0
qrcode~=8.2
reportlab~=4.4.2
requests~=2.32.4
httpx~=0.28.1
postgrest~=1.1.1
python-dotenv~=1.1.1
supabase~=2.16.0